Modified the plotting scripts uu_vs_phi and lu_vs_phi to have a default model value while keeping the option of selecting the model manually using -model. This makes the test plotting example in the README work.
I did not change the cross section plotting script. Only the above scripts were modified.

Opinion: I think these scripts are useful to us in our analysis and presentation but I don't think they belong in the release. We could include a test plot script, with the intent of letting users test their installation. These scripts could be used for that reason as long as they are setup to be completely default.